Rename Application- to Home-Controller and vice versa #14
