+++ app/controllers/HomeController.java
@@ -1,29 +1,34 @@
 package controllers;
 
-import play.inject.Injector;
+import common.cms.CmsPage;
+import common.controllers.ControllerDependency;
+import common.controllers.SunriseController;
+import common.pages.SunrisePageData;
 import play.libs.F;
-import play.mvc.Controller;
 import play.mvc.Result;
-import setupwidget.controllers.SetupController;
+import productcatalog.pages.HomePageContent;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
 /**
- * Front controller for the {@link ApplicationController} to enable a widget to set the SPHERE.IO credentials.
+ * Controller for main web pages like index, imprint and contact.
  */
 @Singleton
-public class HomeController extends Controller {
-    private final Injector injector;
-    private final SetupController setupController;
+public final class HomeController extends SunriseController {
 
     @Inject
-    public HomeController(final Injector injector, final SetupController setupController) {
-        this.injector = injector;
-        this.setupController = setupController;
+    public HomeController(final ControllerDependency controllerDependency) {
+        super(controllerDependency);
     }
 
     public F.Promise<Result> index() {
-        return setupController.handleOrFallback(() -> injector.instanceOf(ApplicationController.class).index());
+        return getCmsPage("home").map(this::getResult);
+    }
+
+    private Result getResult(final CmsPage cms) {
+        final HomePageContent content = new HomePageContent();
+        final SunrisePageData pageData = SunrisePageData.of(cms, context(), content);
+        return ok(templateService().renderToHtml("home", pageData));
     }
 }
