+++ conf/routes
@@ -3,7 +3,7 @@
 # ~~~~
 
 # Home page
-GET        /                                                @controllers.HomeController.index()
+GET        /                                                @controllers.ApplicationController.index()
 
 GET        /products                                        @controllers.TmpPopController.show
